requirements.txt
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
scikit-image   # for LBP
pillow
